==> Notebooks/03_feature_engineering.ipynb <==
import pandas as pd
import os

# load the combined data
combined_data = pd.read_csv('Data/data_v1/grizzlys_combined_schedule_sales_v2.csv')

# print the first few rows to verify
print(combined_data.head())

# print column data types
print(combined_data.dtypes)

# check for missing values
print(combined_data.isnull().sum())

# show rows with missing values
print(combined_data[combined_data.isnull().any(axis=1)])

# show unique values in away_team column
print(combined_data['away_team'].unique())

# replace "Schwenningen" with "Schwenninger Wild Wings" and update the distance accordingly

combined_data.loc[combined_data['away_team'] == 'Schwenningen', 'away_team'] = 'Schwenninger Wild Wings'

# get the distnace from other row where away_team is Schwenninger Wild Wings
distance_value = combined_data.loc[combined_data['away_team'] == 'Schwenninger Wild Wings', 'distance'].iloc[0]
print(f"Distance value to be used: {distance_value}")
# update the missing distance values
combined_data.loc[combined_data['away_team'] == 'Schwenninger Wild Wings', 'distance'] = distance_value

# verify the changes
print(combined_data[combined_data['away_team'] == 'Schwenninger Wild Wings'])

# show rows with missing values
print(combined_data[combined_data.isnull().any(axis=1)])

# update the distance for away_team SC Bietigheim Steelers to 517 km
combined_data.loc[combined_data['away_team'] == 'SC Bietigheim Steelers', 'distance'] = 517

# verify the changes
print(combined_data[combined_data['away_team'] == 'SC Bietigheim Steelers'])

# show rows with missing values
print(combined_data[combined_data.isnull().any(axis=1)])

# drop the rows with missing values
combined_data = combined_data.dropna()

# verify no missing values remain
print(combined_data.isnull().sum())

# save the final cleaned data
output_path = 'Data/data_v1/grizzlys_combined_schedule_sales_v3.csv'
combined_data.to_csv(output_path, index=False)
print(f"Cleaned data saved to {output_path}")

# show unique values in time column
print(combined_data['time'].unique())
